+++ sakura_core/charset/Utf8.h
@@ -57,10 +57,10 @@ public:
 protected:
 	// 変換の実装
 	// 2008.11.10 変換ロジックを書き直す
-	inline static int _Utf8ToUni_char(const unsigned char*, const size_t, unsigned short*, bool bCESU8Mode);
-	static int Utf8ToUni(const char*, const size_t, wchar_t*, bool bCESU8Mode);
-	inline static int _UniToUtf8_char(const unsigned short*, const size_t, unsigned char*, const bool bCSU8Mode);
-	static int UniToUtf8(const wchar_t*, const size_t, char*, bool* pbError, bool bCSU8Mode);
+	inline static size_t _Utf8ToUni_char(const unsigned char*, const size_t, unsigned short*, bool bCESU8Mode);
+	static size_t Utf8ToUni(const char*, const size_t, wchar_t*, bool bCESU8Mode);
+	inline static size_t _UniToUtf8_char(const unsigned short*, const size_t, unsigned char*, const bool bCSU8Mode);
+	static size_t UniToUtf8(const wchar_t*, const size_t, char*, bool* pbError, bool bCSU8Mode);
 };
 
 /*!
@@ -71,9 +71,9 @@ protected:
 	高速化のため、インライン化
 
 */
-inline int Utf8::_Utf8ToUni_char( const unsigned char* pSrc, const size_t nSrcLen, unsigned short* pDst, bool bCESUMode )
+inline size_t Utf8::_Utf8ToUni_char( const unsigned char* pSrc, const size_t nSrcLen, unsigned short* pDst, bool bCESUMode )
 {
-	int nret;
+	size_t nret;
 
 	if (nSrcLen < 1) {
 		return 0;
@@ -123,9 +123,9 @@ inline int Utf8::_Utf8ToUni_char( const unsigned char* pSrc, const size_t nSrcLe
 
 	高速化のため、インライン化
 */
-inline int Utf8::_UniToUtf8_char( const unsigned short* pSrc, const size_t nSrcLen, unsigned char* pDst, bool bCESU8Mode )
+inline size_t Utf8::_UniToUtf8_char( const unsigned short* pSrc, const size_t nSrcLen, unsigned char* pDst, bool bCESU8Mode )
 {
-	int nret;
+	size_t nret;
 
 	if (nSrcLen < 1) {
 		return 0;
